functional recorder
At this stage, it's good enough that I will use it for myself. There's still features missing. Here's a list:
-1 Deleting & Updating trainee records
-2 Creating spark records
-3 Linking the records for races/skills/cards/sparks to the veterans
-4 The ability to add Guest Trainees
-5 The ability to link trainees to eachother based on who's inspired who

If I added all those, then it would be a conclusive version with everything I need to record *all* of a Veteran's data, and comfortably delete them from my ingame account.

-6 A system for creating database backups.
This is needed because, when the website is exclusively accessible locally, there's nothing stopping me from manually executing reckless SQL commands. I like doing it, and I'm the one who will suffer from it when my data is deleted.

-7 a php file which tests my DELETE and UPDATE sql commands before allowing me to do them. By performing the SELECT portion on the relevant records and showing me how they'll be changed, for example.